Providing user specified schema.
